requirements.txt
click~=7.1.2
paho-mqtt~=1.5.1
setuptools~=53.0.0